END OF TARGIL 2

+++ dotNet5782_8995_8035/BL/DroneInCharge.cs
@@ -11,7 +11,7 @@ namespace IBAL
         public class DroneInCharge
         {
             public int Id { get; set; }
-            public int Battary { get; set; }
+            public double Battary { get; set; }
 
             public override string ToString()
             {
